Add function to create topic json from the various json pieces.
Adds a function (createTopicJson) to create the topic json from the
various json definitions, theorems and extra code.
Also change how lists and binary trees are definied, making them
polymorphic, so we don't have to duplicate them for every type we use
them on (this needed a bit of change in the TeXifier to make them show
properly).